# schemas/openapi.v1.json
{
  "components": {
    "schemas": {
...
      "AlignmentStatus": {
        "enum": [
          "unaligned",
          "approximate",
          "aligned",
          "reviewed"
        ],
        "title": "AlignmentStatus",
        "type": "string"
      },
      "ArtifactDetail": {
        "properties": {
          "alignment_status": {
            "$ref": "#/components/schemas/AlignmentStatus"
          },
          "artifact_id": {
            "title": "Artifact Id",
            "type": "string"
          },
          "capture_id": {
            "title": "Capture Id",
            "type": "string"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "footprint": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/GeoJSONPolygon"
              },
              {
                "type": "null"
              }
            ]
          },
          "frame_name": {
            "title": "Frame Name",
            "type": "string"
          },
          "kind": {
            "title": "Kind",
            "type": "string"
          },
          "layer_default": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/LayerDefault"
              },
              {
                "type": "null"
              }
            ]
          },
          "manifest_sha256": {
            "title": "Manifest Sha256",
            "type": "string"
          },
          "point_count": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Point Count"
          },
          "representations": {
            "items": {
              "$ref": "#/components/schemas/RepresentationResponse"
            },
            "title": "Representations",
            "type": "array"
          },
          "run_id": {
            "title": "Run Id",
            "type": "string"
          },
          "units": {
            "title": "Units",
            "type": "string"
          }
        },
        "required": [
          "artifact_id",
          "run_id",
          "capture_id",
          "kind",
          "alignment_status",
          "frame_name",
          "units",
          "created_at",
          "manifest_sha256",
          "representations"
        ],
        "title": "ArtifactDetail",
        "type": "object"
      },
...
      "CaptureResponse": {
        "properties": {
          "capture_id": {
            "title": "Capture Id",
            "type": "string"
          },
          "device": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Device"
          },
          "ended_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ended At"
          },
          "footprint": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/GeoJSONPolygon"
              },
              {
                "type": "null"
              }
            ]
          },
          "fps": {
            "anyOf": [
              {
                "exclusiveMinimum": 0.0,
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Fps"
          },
          "frame_count": {
            "anyOf": [
              {
                "minimum": 0.0,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Frame Count"
          },
          "lens": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Lens"
          },
          "source_uri": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Source Uri"
          },
          "started_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Started At"
          },
          "video_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Video Name"
          }
        },
        "required": [
          "capture_id"
        ],
        "title": "CaptureResponse",
        "type": "object"
      },
      "CatalogArtifactsResponse": {
        "properties": {
          "artifacts": {
            "items": {
              "$ref": "#/components/schemas/ArtifactSummary"
            },
            "title": "Artifacts",
            "type": "array"
          },
          "count": {
            "minimum": 0.0,
            "title": "Count",
            "type": "integer"
          }
        },
        "required": [
          "count",
          "artifacts"
        ],
        "title": "CatalogArtifactsResponse",
        "type": "object"
      },
...
      "ErrorResponse": {
        "properties": {
          "detail": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "items": {
                  "additionalProperties": true,
                  "type": "object"
                },
                "type": "array"
              }
            ],
            "title": "Detail"
          }
        },
        "required": [
          "detail"
        ],
        "title": "ErrorResponse",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "Manifest": {
        "additionalProperties": false,
        "properties": {
          "alignment": {
            "$ref": "#/components/schemas/Alignment"
          },
          "artifact_id": {
            "maxLength": 200,
            "minLength": 1,
            "title": "Artifact Id",
            "type": "string"
          },
          "artifacts": {
            "items": {
              "$ref": "#/components/schemas/ArtifactFile"
            },
            "minItems": 1,
            "title": "Artifacts",
            "type": "array"
          },
          "capture": {
            "$ref": "#/components/schemas/CaptureMetadata"
          },
          "capture_id": {
            "maxLength": 200,
            "minLength": 1,
            "title": "Capture Id",
            "type": "string"
          },
          "coordinate_frame": {
            "$ref": "#/components/schemas/CoordinateFrame"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "footprint_wgs84": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/WGS84Footprint"
              },
              {
                "type": "null"
              }
            ]
          },
          "layer_default": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/LayerDefault"
              },
              {
                "type": "null"
              }
            ]
          },
          "lossy_operations": {
            "items": {
              "$ref": "#/components/schemas/LossyOperation"
            },
            "title": "Lossy Operations",
            "type": "array"
          },
          "producer": {
            "$ref": "#/components/schemas/Producer"
          },
          "run_id": {
            "maxLength": 200,
            "minLength": 1,
            "title": "Run Id",
            "type": "string"
          },
          "schema_version": {
            "const": "1.0.0",
            "default": "1.0.0",
            "title": "Schema Version",
            "type": "string"
          }
        },
        "required": [
          "run_id",
          "capture_id",
          "artifact_id",
          "created_at",
          "capture",
          "producer",
          "coordinate_frame",
          "alignment",
          "artifacts"
        ],
        "title": "Manifest",
        "type": "object"
      },
...
      "RunResponse": {
        "properties": {
          "capture_id": {
            "title": "Capture Id",
            "type": "string"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "git_commit": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Git Commit"
          },
          "git_status": {
            "enum": [
              "clean",
              "dirty",
              "unknown"
            ],
            "title": "Git Status",
            "type": "string"
          },
          "manifest_sha256": {
            "title": "Manifest Sha256",
            "type": "string"
          },
          "model_name": {
            "title": "Model Name",
            "type": "string"
          },
          "model_version": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Model Version"
          },
          "run_id": {
            "title": "Run Id",
            "type": "string"
          },
          "status": {
            "$ref": "#/components/schemas/AlignmentStatus"
          }
        },
        "required": [
          "run_id",
          "capture_id",
          "status",
          "created_at",
          "model_name",
          "git_status",
          "manifest_sha256"
        ],
        "title": "RunResponse",
        "type": "object"
      },
...
      "SourcesResponse": {
        "properties": {
          "count": {
            "minimum": 0.0,
            "title": "Count",
            "type": "integer"
          },
          "sources": {
            "items": {
              "$ref": "#/components/schemas/SourceRecord"
            },
            "title": "Sources",
            "type": "array"
          }
        },
        "required": [
          "count",
          "sources"
        ],
        "title": "SourcesResponse",
        "type": "object"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    }
  },
  "info": {
    "description": "Canonical reconstruction catalog and immutable artifact server",
    "title": "Mapper Catalog API",
    "version": "1.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
    "/api/v1/assets/{representation_id}": {
      "get": {
        "operationId": "get_asset_api_v1_assets__representation_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "representation_id",
            "required": true,
            "schema": {
              "title": "Representation Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "206": {
            "description": "Requested representation byte range"
          },
          "304": {
            "description": "Representation ETag matched"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Conflict"
          },
          "416": {
            "description": "Requested range is not satisfiable"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Asset"
      },
      "head": {
        "operationId": "head_asset_api_v1_assets__representation_id__head",
        "parameters": [
          {
            "in": "path",
            "name": "representation_id",
            "required": true,
            "schema": {
              "title": "Representation Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Head Asset"
      }
    },
    "/api/v1/catalog/artifacts": {
      "get": {
        "operationId": "query_artifacts_api_v1_catalog_artifacts_get",
        "parameters": [
          {
            "in": "query",
            "name": "bbox",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Bbox"
            }
          },
          {
            "in": "query",
            "name": "started_after",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "date-time",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Started After"
            }
          },
          {
            "in": "query",
            "name": "ended_before",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "date-time",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Ended Before"
            }
          },
          {
            "in": "query",
            "name": "alignment_status",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/AlignmentStatus"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Alignment Status"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 100,
              "maximum": 1000,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CatalogArtifactsResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unprocessable Entity"
          }
        },
        "summary": "Query Artifacts"
      }
    },
    "/api/v1/catalog/artifacts/{artifact_id}": {
      "get": {
        "operationId": "get_artifact_api_v1_catalog_artifacts__artifact_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "artifact_id",
            "required": true,
            "schema": {
              "title": "Artifact Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ArtifactDetail"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not Found"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Artifact"
      }
    },
    "/api/v1/catalog/artifacts/{artifact_id}/manifest": {
      "get": {
        "operationId": "get_manifest_api_v1_catalog_artifacts__artifact_id__manifest_get",
        "parameters": [
          {
            "in": "path",
            "name": "artifact_id",
            "required": true,
            "schema": {
              "title": "Artifact Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Manifest"
                }
              }
            },
            "description": "Successful Response"
          },
          "206": {
            "description": "Requested manifest byte range"
          },
          "304": {
            "description": "Manifest ETag matched"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not Found"
          },
          "416": {
            "description": "Requested range is not satisfiable"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Manifest"
      },
      "head": {
        "operationId": "head_manifest_api_v1_catalog_artifacts__artifact_id__manifest_head",
        "parameters": [
          {
            "in": "path",
            "name": "artifact_id",
            "required": true,
            "schema": {
              "title": "Artifact Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "206": {
            "description": "Requested manifest byte range"
          },
          "304": {
            "description": "Manifest ETag matched"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not Found"
          },
          "416": {
            "description": "Requested range is not satisfiable"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Head Manifest"
      }
    },
    "/api/v1/catalog/artifacts/{artifact_id}/sources": {
      "get": {
        "operationId": "get_sources_api_v1_catalog_artifacts__artifact_id__sources_get",
        "parameters": [
          {
            "in": "path",
            "name": "artifact_id",
            "required": true,
            "schema": {
              "title": "Artifact Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SourcesResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not Found"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Sources"
      }
    },
    "/api/v1/catalog/captures/{capture_id}": {
      "get": {
        "operationId": "get_capture_api_v1_catalog_captures__capture_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "capture_id",
            "required": true,
            "schema": {
              "title": "Capture Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CaptureResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not Found"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Capture"
      }
    },
    "/api/v1/catalog/runs/{run_id}": {
      "get": {
        "operationId": "get_run_api_v1_catalog_runs__run_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "run_id",
            "required": true,
            "schema": {
              "title": "Run Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RunResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not Found"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Run"
      }
    },
...
  }
}